+++ src/collective/metadataversion/utils.py
@@ -92,7 +92,8 @@ def _update_mmu_kwargs(context, logger, metadata_version, kwargs):
     comparisons as intended; but it won't be stored by default:
     >>> kw = {}
     >>> ummu(41, kw)                          # doctest: +NORMALIZE_WHITESPACE
-    [('idxs', ['getId']),
+    [('change',        None),
+     ('idxs',         ['getId']),
      ('minimum_version', 41),
      ('new_version',   None),
      ('old_version',     42)]
@@ -100,7 +101,8 @@ def _update_mmu_kwargs(context, logger, metadata_version, kwargs):
     ... unless we force it to be accepted:
     >>> kw = dict(force_version=1)
     >>> ummu(41, kw)                          # doctest: +NORMALIZE_WHITESPACE
-    [('idxs', ['getId']),
+    [('change',        None),
+     ('idxs',         ['getId']),
      ('minimum_version', 41),
      ('new_version',     41),
      ('old_version',     42)]
@@ -109,7 +111,8 @@ def _update_mmu_kwargs(context, logger, metadata_version, kwargs):
     will result in:
     >>> kw = {}
     >>> ummu(None, kw)                        # doctest: +NORMALIZE_WHITESPACE
-    [('idxs', ['getId']),
+    [('change',        None),
+     ('idxs',         ['getId']),
      ('minimum_version', 41),
      ('new_version',   None),
      ('old_version',     41)]
@@ -139,6 +142,7 @@ def _update_mmu_kwargs(context, logger, metadata_version, kwargs):
                               'we expect a non-None metadata_version!')
     unused = set(kwargs) - set(['idxs', 'force_indexes',
                                 'update_metadata',
+                                'change',  # a function(object, logger)
                                 'debug',
                                 ])
     if unused:
@@ -173,6 +177,7 @@ def _update_mmu_kwargs(context, logger, metadata_version, kwargs):
             idxs = list(registry[FULL_IDXS_KEY])  # default: a cheap subset
         kwargs['idxs'] = idxs
 
+    kwargs.setdefault('change', None)  # a function(object, logger)
     kwargs.update({
         # metadata_version value, stored in the registry:
         'new_version':     new_version,  # if not None, written to registry
@@ -225,6 +230,17 @@ def make_metadata_updater(context, logger=None, metadata_version=None,
     Thus, if you *specify* ``idxs=None`` but not `force_indexes`, all indexes
     will be updated as well if the metadata is considered outdated.
 
+    change -- a function(object, logger) to be called before reindexing is
+              performed.
+              As what we do depends on the metadata_version found in the
+              catalog object, this function call does so as well:
+              If our negotiation concludes to not reindex, this function won't
+              be called, either.
+              Currently the returned value is ignored;
+              if the changes applied by the function might cause more idxs to
+              rot, you currently need to add those to the list (for all),
+              or to specify idxs=None (which will include all indexes).
+
     ... and finally:
 
     update_metadata -- Refresh the metadata columns?
@@ -275,6 +291,15 @@ def make_metadata_updater(context, logger=None, metadata_version=None,
         _info.append('DEBUG')
 
     logger.info('make_metadata_updater: ' + '; '.join(_info))
+    change = kwargs['change']
+    if change is None:
+        pass
+    elif not callable(change):
+        raise ValueError("change option is expected to be a function"
+                '(object, logger); found %s' % (type(change),))
+    else:
+        logger.info('The %(change)r function(object, logger) will be '
+                    'called before reindexing an object', locals())
 
     def reindex(brain):
         """
@@ -317,8 +342,43 @@ def make_metadata_updater(context, logger=None, metadata_version=None,
                 raise ObjectNotFound(txt, e)
 
         try:
+            if change is not None:
+                res = change(o, logger)
+                if res is not None:
+                    logger.debug('%(res)r <-- change(%(o)r, ...)', locals())
+                    # we might e.g. inject indexes to be updated ...
+                    logger.warn('change(%(o)r: Currently, the returned value '
+                                'is ignored!', locals())
             catalog_reindex(o, idxs=idxs, update_metadata=refresh_metadata)
-        except (ConflictError, KeyboardInterrupt):
+        except IOError as e:
+            # There is a likely source for IOError exception:
+            # PIL (or Pillow), objecting to an unsupported image format.
+            # Of course such errors should be caught by the respective indexer,
+            # allowing for the rest of the objects to be reindexed;
+            # but for now, we like to avoid the whole job to fail
+            # just because of very few foul objects.
+            msg = e.message
+            fail = 1
+            cause = ''
+            if msg is not None:
+                if (msg == 'unrecognized data stream contents'
+                           ' when reading image file'
+                    or msg.startswith('cannot find loader for this ')
+                    ):
+                    cause = ' (PIL)'
+                    fail = 0
+            logger.error('IOError reindexing %(o)r%(cause)s: %(e)r', locals())
+            if fail:
+                raise ReindexingError('Error reindexing %(o)r' % locals(),
+                                      e)
+            else:
+                return False
+        except ConflictError as e:
+            logger.error('ConflictError reindexing %(o)r: %(e)r', locals())
+            # ConflictErrors must never be masked, right?
+            raise
+        except KeyboardInterrupt:
+            logger.error('INTERRUPTED reindexing %(o)r!', locals())
             raise
         except Exception as e:
             logger.error('error reindexing %(o)r: %(e)r', locals())
@@ -372,6 +432,7 @@ def extract_mmu_kwargs(kw, do_pop=1):
         'metadata_version',
         'idxs',
         'force_version', 'force_indexes',
+        'change',  # function(object, logger)
         ):
         if key in kw:
             res[key] = get(key)